+++ MSAL/src/public/MSAL.h
@@ -88,4 +88,5 @@ FOUNDATION_EXPORT const unsigned char MSAL__Framework_VersionString[];
 #import <MSAL/MSALAuthenticationSchemePop.h>
 #import <MSAL/MSALAuthenticationSchemeProtocol.h>
 #import <MSAL/MSALHttpMethod.h>
+#import <MSAL/MSALCIAMAuthority.h>
 #import <MSAL/MSALWipeCacheForAllAccountsConfig.h>